[Mellanox] build mft/hw-management in buster stage
Currently we port SONiC to buster in a way that base image is on buster and
other dockers based on stretch. The benefit is that tasks can be carried out
simultaneously.

The build procedure can be treated as 2 stages.

The first stage is to build the stretch-based debs and dockers and the second
stage is to build the buster-based ones.

One thing we have to pay attention to is some debs depend on kernel should not
be built at stretch stage because the kernel isn't available at that time.
The idea is to move that kind of debs out of SONIC_STRETCH_DEBS. Meanwhile,
any dependency explicitly put on the stretch based dockers on kernel should be
removed.